+++ cookbooks/aws-parallelcluster-shared/resources/install_pyenv.rb
@@ -16,11 +16,12 @@ action :run do
   python_version = new_resource.python_version || node['cluster']['python-version']
 
   if new_resource.user_only
+    # User install seems to be broken, but is only used in scheduler plugin
     raise "user property is required for resource install_pyenv when user_only is set to true" unless new_resource.user
 
-    pyenv_user_install python_version do
+    pyenv_install 'user' do
       user new_resource.user
-      user_prefix new_resource.prefix if new_resource.prefix
+      prefix new_resource.prefix if new_resource.prefix
     end
   else
     prefix = new_resource.prefix || node['cluster']['system_pyenv_root']
@@ -29,8 +30,8 @@ action :run do
       recursive true
     end
 
-    pyenv_system_install python_version do
-      global_prefix prefix
+    pyenv_install 'system' do
+      prefix prefix
     end
 
     # Remove the profile.d script that the pyenv cookbook writes.